Data dump files for Clock support configuration (#188)
<details>
  <summary>  Keywords used:</summary>


clocksource
chrony
timer
adjtime
timekeeping
jiffies
CLOCK_REALTIME
CLOCK_MONOTONIC
hrtimer
timecounter
tick
clockevents
timeconst


[jiffies](https://elixir.bootlin.com/linux/latest/source/include/linux/jiffies.h)
: It is a global variable that holds the number of ticks that have
occured since the system booted.
Its intial value is set to zero, incremented by one during each time
interrupt.
Declared inside <linux/jiffies.h> 


[clocksource](https://www.kernel.org/doc/Documentation/timers/timekeeping.txt)
: Provides a timeline for the system that tells one where it is in time.


[timekeeping](https://www.kernel.org/doc/Documentation/timers/timekeeping.txt)
: Used for tracking and handling system timers.
Ensures that the system clock is synchronized with external time sources
and provide accurate time info to user-space applications.

[timer](https://elinux.org/Kernel_Timer_Systems) : A mechanism, that
allows a specific block of code to execute after a specific amount of
time has elapsed.

chrony: Implementation of the Network Time Protocol(NTP). Used for
synchronization of the system clock of a computer with external time
reference.


[adjtime](https://elixir.bootlin.com/linux/latest/source/include/linux/ptp_clock_kernel.h#L172)
: System call that is used to adjust the time offset of the system call.
Adjust the frequency of the clock oscillator.


[hrtimer](https://elixir.bootlin.com/linux/latest/source/kernel/time/hrtimer.c)
: It is a high resolution timer framemwork. Used for scheduling tasks in
the kernel with nanosecond accuracy


[CLOCK_MONOTONIC](https://www.google.com/url?sa=t&rct=j&q=&esrc=s&source=web&cd=&ved=2ahUKEwiz5rGk1rj9AhW-7TgGHe9sB_EQFnoECAwQAQ&url=https%3A%2F%2Felixir.bootlin.com%2Flinux%2Flatest%2Fident%2FCLOCK_MONOTONIC&usg=AOvVaw0sE_0mWJYGm7jnGaFUcvSl)
: clock that provides a monotonically increasing time value, meaning the
time value will never decrease, and is not affected by system time
changes. It is commonly used for measuring elapsed time or timeouts in
programs and is considered to be the most accurate time source available
on Linux systems.


[CLOCK_REALTIME](https://elixir.bootlin.com/linux/v4.5/ident/CLOCK_REALTIME)
: It is a clock source that represents the wall-clock time. It is a
system-wide clock that measures the time elapsed. It is generally used
to measure the real-world time of events, and is commonly used in
user-space applications for scheduling and synchronization.

[clock_gettime](https://linux.die.net/man/3/clock_gettime) : It is a
system call in the Linux kernel that retrieves the current time of the
system clock specified by the caller.

clocksource_driver : It is a kernel subsystem that provides an interface
for device drivers to register a clocksource with the kernel. It manages
the clocksource devices and provides a consistent timekeeping interface
to the rest of the kernel.

timecounter : It is a mechanism to keep count of the time.
</details>