Script for Polling Data Dump & Resource Manipulation (#198)
### Details about some of the keywords:


[Poll](https://elixir.bootlin.com/linux/v3.7/source/include/linux/poll.h)
: It a system call used to monitor multiple file descriptors for events.

Polling : Polling is a technique used to check for the status of events
or conditions in the kernel.

[select](https://elixir.bootlin.com/linux/v2.6.11/source/fs/select.c) :
It is a system call that monitors multiple file descriptors, waiting
until one or more of the file descriptors become ready for some kind of
I/O operation.

[epoll](https://man7.org/linux/man-pages/man7/epoll.7.html) : It is an
I/O event notification facility in the Linux kernel that replaces the
older select and poll system calls.

[Interrupt Handlers]() : Thsese are the kernel routines that handle the
interrupts generated by hardware devices.
Related Keyword: irq


[Waitqueue](https://elixir.bootlin.com/linux/v6.0.16/source/include/linux/wait.h#L37)
: Ths used for synchronization between different parts of the kernel,
such as a process waiting for an event to occur or for a resource to
become available. When a process wants to wait for an evnt, it can ad
itself to a wait queue associated with that event. When the event
occurs, the kernel wakes up all the processes that are waiting on the
queue.
Related Data structure: wait_queue_head_t


[Interrupts](https://elixir.bootlin.com/linux/v6.0.16/source/include/linux/interrupt.h)
: These are the signals sent by hardware devices to the CPU to request
its attention.
Related Keyword: irq_desc, irq_chip, net/core/dev.c, block/blk-core.c


[Tasklet](https://github.com/torvalds/linux/blob/master/kernel/softirq.c)
: A tasklet is a kernel mechanism used to schedule a function to run
later in a low-level interrupt context.
Related Keyword: tasklet_init


[Workqueue](https://elixir.bootlin.com/linux/v6.0.16/source/kernel/workqueue.c)
: It is a kernel mechanism that allows work to be deferred to a kernel
thread, running in the background.