Script for Hardware Tracing Log (#204)
Related Content:

**CoreSight** is a debugging and tracing technology developed by ARM for
use in their processors. It provides a range of hardware features that
enable real-time debugging, trace, and performance monitoring of
embedded systems.

**TPIU** (Trace Port Interface Unit): It provides a standard interface
for trace data between the SoC and a trace capture device. It enables
real-time trace data capture and analysis for debugging and system
optimization.

**ETB** (Embedded Trace Buffer): It provides a trace buffer for storing
trace data from the processor. It is used to capture trace data from the
processor and then store it in a buffer for later analysis.

**STM** (System Trace Macrocell): It provides a mechanism for capturing
and storing system-wide trace data. It can capture trace data from
multiple sources, such as CPUs, memory buses, and peripheral buses.

**PTM** (Program Trace Macrocell): Module that provides a mechanism for
tracing program execution.

**CTI** (Cross Trigger Interface): It is a hardware module within the
CoreSight infrastructure that provides a mechanism for triggering events
and passing data between different trace sources.

**TMC** (Trace Memory Controller): A module within the CoreSight that
provides a mechanism for controlling access to the trace memory.

**ETF** (Embedded Trace FIFO): Another module in CoreSight that provides
a trace buffer for storing trace data from the processor.